fix itinerary tab bug and optimize layout
