+++ src/pages/UserPage.jsx
@@ -1,12 +1,10 @@
 import { Box } from "@mui/material";
-import LabTabs from "../components/UserPage/LabTab";
+import LabTab from "../components/UserPage/LabTab";
 import CreateTripModal from "../components/UserPage/CreateTripModel.jsx";
 import Navbar from "../components/Home/Navbar.jsx";
-import { ItinerariesProvider } from "../context/UserPageContext.jsx";
+import { ItinerariesProvider, useItineraries } from "../context/UserPageContext.jsx";
 import { useEffect, useState } from "react";
-import { getItineraries, getJoinedItinerariesId } from "../api/userpage.jsx";
-import { useItineraries } from "../context/UserPageContext.jsx";
-import { getItinerary } from "../api/userpage.jsx";
+import { getItineraries, getJoinedItinerariesId, getItinerary } from "../api/userpage.jsx";
 
 export default function UserPage() {
   return (
@@ -29,15 +27,17 @@ function UserPageContent() {
   }, [])
 
   // fetch total participated itineraries id when first render
+  // use itineraries as dependency in order to rerender page whenever add or delete itinerary
   useEffect(()=>{
     getJoinedItinerariesId()
     .then(data => {
       setId(data)
       // console.log(data)
     })
-  }, [])
+  }, [itineraries])
 
   // fetch all participated itineraries data
+  // use id as dependency in order to rerender page whenever new participated id generated
   useEffect(() => {
     const fetchItineraries = async () => {
       const itinerariesData = await Promise.all(id.map(id => getItinerary(id)));
@@ -55,7 +55,7 @@ function UserPageContent() {
           <Box sx={{ flexGrow: 1 }} bgcolor='white' >
             <Box width="100%" height="100%" sx={{ background: "white", p: 2 }}>
               <h1 style={{color:'#38358C', marginTop:'0.5vw',marginBottom:'1.5vw', fontFamily:'Poppins', fontWeight:600, fontSize:40 }}>My Account</h1>
-              <LabTabs image={'../../src/images/spot/California.jpeg'} />
+              <LabTab />
               <CreateTripModal sx={{ position:'fixed', bottom:30, right:30 }} />
             </Box>
           </Box>
